update the type

+++ include/aws/s3/private/s3_auto_ranged_put.h
@@ -107,7 +107,7 @@ AWS_EXTERN_C_BEGIN
 AWS_S3_API struct aws_s3_meta_request *aws_s3_meta_request_auto_ranged_put_new(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    size_t part_size,
+    uint64_t part_size,
     uint64_t content_length,
     uint32_t num_parts,
     const struct aws_s3_meta_request_options *options);
