+++ include/aws/s3/private/s3_util.h
@@ -232,7 +232,7 @@ int aws_s3_parse_content_length_response_header(
  * part-ranges on part_size. (ie: if object_range_start is not evenly divisible by part_size, it is considered in the
  * middle of a contiguous part, and that first part will be smaller than part_size.) */
 AWS_S3_API
-uint32_t aws_s3_get_num_parts(size_t part_size, uint64_t object_range_start, uint64_t object_range_end);
+uint32_t aws_s3_get_num_parts(uint64_t part_size, uint64_t object_range_start, uint64_t object_range_end);
 
 /* Calculates the part range for a part given overall object range, size of each part, and the part's number. Note: part
  * numbers begin at one. This takes into account aligning part-ranges on part_size. Intended to be used in conjunction
@@ -241,7 +241,7 @@ AWS_S3_API
 void aws_s3_get_part_range(
     uint64_t object_range_start,
     uint64_t object_range_end,
-    size_t part_size,
+    uint64_t part_size,
     uint32_t part_number,
     uint64_t *out_part_range_start,
     uint64_t *out_part_range_end);
